[MOB-72] Moved download and download persistence classes to correct packages
