+++ app/src/main/java/cm/aptoide/pt/wallet/WalletInstallAnalytics.kt
@@ -6,7 +6,7 @@ import cm.aptoide.analytics.implementation.navigation.ScreenTagHistory
 import cm.aptoide.pt.ads.WalletAdsOfferManager
 import cm.aptoide.pt.app.DownloadModel
 import cm.aptoide.pt.app.DownloadStateParser
-import cm.aptoide.pt.database.realm.RoomDownload
+import cm.aptoide.pt.database.room.RoomDownload
 import cm.aptoide.pt.download.AppContext
 import cm.aptoide.pt.download.DownloadAnalytics
 import cm.aptoide.pt.install.InstallAnalytics
