+++ app/src/vanillaDev/java/cm/aptoide/pt/install/PackageInstallerManager.java
@@ -1,6 +1,6 @@
 package cm.aptoide.pt.install;
 
-import cm.aptoide.pt.database.realm.RoomDownload;
+import cm.aptoide.pt.database.room.RoomDownload;
 
 public class PackageInstallerManager {
   boolean shouldSetInstallerPackageName(RoomDownload download) {
