+++ downloadmanager/src/main/java/cm/aptoide/pt/downloadmanager/DownloadAnalytics.java
@@ -1,6 +1,6 @@
 package cm.aptoide.pt.downloadmanager;
 
-import cm.aptoide.pt.database.realm.RoomDownload;
+import cm.aptoide.pt.database.room.RoomDownload;
 
 public interface DownloadAnalytics {
