+++ downloadmanager/src/main/java/cm/aptoide/pt/downloadmanager/DownloadsRepository.java
@@ -1,7 +1,7 @@
 package cm.aptoide.pt.downloadmanager;
 
 import cm.aptoide.pt.database.accessors.RoomDownloadPersistence;
-import cm.aptoide.pt.database.realm.RoomDownload;
+import cm.aptoide.pt.database.room.RoomDownload;
 import java.util.List;
 import rx.Completable;
 import rx.Observable;
